move AdditionalInputs model to dto package
