+++ src/backend/domain/dto/courses/course_plan.py
@@ -2,6 +2,28 @@ from typing import Optional
 from pydantic import BaseModel
 from domain.enums.course_enums import CourseMotivation, CurrentSubjectExperience, CourseMaterial
 
+from enum import Enum
+from typing import Optional
+from pydantic import BaseModel
+
+class InputType(Enum):
+    TEXT = "text"
+    SELECT = "select"
+    MULTISELECT = "multiselect"
+
+class InputField(BaseModel):
+    required: bool
+    name: str
+    short_label: str
+    description: Optional[str] = None
+    input_type: InputType
+    options: Optional[list[str]] = None
+    depends_on_input_name: Optional[str] = None
+    depends_on_input_value: Optional[str] = None
+
+class AdditionalInputs(BaseModel):
+    inputs: list[InputField]
+
 class CoursePlanDto(BaseModel):
     subject: str
     motivations: list[CourseMotivation]
@@ -9,3 +31,4 @@ class CoursePlanDto(BaseModel):
     experience: CurrentSubjectExperience
     experience_details: Optional[str] = None
     materials: list[CourseMaterial]
+    follow_up_questions: Optional[list[AdditionalInputs]] = None
